Add scheduler
PPU rework in-progress

+++ include/config.h
@@ -8,10 +8,10 @@
 static const int MASTER_CLOCK_FREQ = 4194304;
 static const float MS_PER_FRAME = 16.74;
 static const float FRAME_RATE = 59.7275;
-static const int CYCLE_PER_FRAME  = 17556;
+static const uint16_t CYCLE_PER_FRAME  = 17556;
 
 // APU CONST
 static const uint16_t SAMPLE_COUNT = 2048;
-static const int OUTPUT_FREQUENCY = 44100;
+static const uint16_t OUTPUT_FREQUENCY = 44100;
 
 #endif //CONFIG_H
